Separate out XF base module and individual XF devices to separate files
